add master boot record support
When booting, the bootloader should only wait for
a new image when either no image has been flashed yet or
the user explicitly wishes to update the image. Another reason
would be when a previous flash failed (image corruped, image doesn't
reach the main loop).

Apart from that, the bootloader should immidiately start the application.

The user can now invoke the extended delay by the already established
ECMD command "bootloader".

+++ protocols/uip/ipv6.c
@@ -55,7 +55,7 @@ extern struct uip_eth_addr uip_ethaddr;
 #define uip_icmp6chksum() (upper_layer_chksum(UIP_PROTO_ICMP6))
 
 #ifdef BOOTLOADER_SUPPORT
-extern uint8_t bootload_delay;
+extern uint16_t bootload_delay;
 #endif
 
 #if UIP_CONF_IPV6 && defined(ETHERNET_SUPPORT)
